Fix earlier commit for 64-bit.
git-svn-id: svn://svn.code.sf.net/p/polyml/code/trunk/polyml@1223 ae7d391e-3f74-4a2b-a0db-a8776840fd2a

+++ libpolyml/memmgr.cpp
@@ -125,7 +125,7 @@ LocalMemSpace* MemMgr::NewLocalSpace(POLYUNSIGNED size, bool mut)
         // reserved space.  This ensures that this much space will always
         // be available for C stacks and the C++ heap.
         void *reservation = 0;
-        unsigned int rSpace = reservedSpace*sizeof(PolyWord);
+        size_t rSpace = reservedSpace*sizeof(PolyWord);
 
         if (reservedSpace != 0) {
             reservation = osMemoryManager->Allocate(rSpace, PERMISSION_READ);
